+++ lwk_wasm/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* removed `PsetBalance::fee()`, replacement `PsetBalance::fees_in(asset_id)` ([ELIP-0204](https://github.com/ElementsProject/ELIPs/blob/main/elip-0204.mediawiki)).
+
 ## 0.17.2
 
 * `Network::to_string()` changed output for regtest variant from "ElementsRegtest { policy_asset: <32 byte hex>}" to "CustomElements(ElementsParams { policy_asset: <32 byte hex>, genesis_hash: <32 byte hex> })".